Refactor into separated classes for clarity sake
